Fix a bunch of individual warnings

+++ src/shared/Log.cpp
@@ -881,7 +881,7 @@ void Log::outWorldPacketDump(uint32 socket, uint32 opcode, char const* opcodeNam
 
     fprintf(worldLogfile, "\n%s:\nSOCKET: %u\nLENGTH: %u\nOPCODE: %s (0x%.4X)\nDATA:\n",
             incoming ? "CLIENT" : "SERVER",
-            socket, packet->size(), opcodeName, opcode);
+            socket, static_cast<uint32>(packet->size()), opcodeName, opcode);
 
     size_t p = 0;
     while (p < packet->size())
